+++ src/xz/writer.rs
@@ -12,7 +12,7 @@ use crate::{
     filter::{FilterConfig, FilterType, bcj::BcjWriter, delta::DeltaWriter},
 };
 
-enum FilterWriter<W: Write> {
+pub(super) enum FilterWriter<W: Write> {
     Counting(CountingWriter<W>),
     Lzma2(Box<Lzma2Writer<Box<FilterWriter<W>>>>),
     Delta(Box<DeltaWriter<Box<FilterWriter<W>>>>),
@@ -55,7 +55,7 @@ impl<W: Write> FilterWriter<W> {
         FilterWriter::Bcj(Box::new(writer))
     }
 
-    fn create_filter_chain(
+    pub(super) fn create_filter_chain(
         inner: CountingWriter<W>,
         filters: &[FilterConfig],
         lzma_options: &LzmaOptions,
@@ -173,7 +173,7 @@ impl<W: Write> FilterWriter<W> {
         }
     }
 
-    fn finish(self) -> Result<CountingWriter<W>> {
+    pub(super) fn finish(self) -> Result<CountingWriter<W>> {
         match self {
             FilterWriter::Counting(writer) => Ok(writer),
             FilterWriter::Lzma2(writer) => {